Fix review comment: Move to 'audit' packages
