+++ integration-tests/src/test/java/com/kenshoo/pl/audit/AuditForDeleteOneLevelTest.java
@@ -7,9 +7,10 @@ import com.kenshoo.pl.audit.commands.DeleteAuditedCommand;
 import com.kenshoo.pl.audit.commands.DeleteAuditedWithoutDataFieldsCommand;
 import com.kenshoo.pl.audit.commands.DeleteNotAuditedCommand;
 import com.kenshoo.pl.entity.*;
+import com.kenshoo.pl.entity.audit.AuditRecord;
+import com.kenshoo.pl.entity.internal.audit.MainTable;
 import com.kenshoo.pl.entity.internal.audit.entitytypes.AuditedType;
 import com.kenshoo.pl.entity.internal.audit.entitytypes.AuditedWithoutDataFieldsType;
-import com.kenshoo.pl.entity.internal.audit.MainTable;
 import com.kenshoo.pl.entity.internal.audit.entitytypes.NotAuditedType;
 import org.jooq.DSLContext;
 import org.junit.After;
