+++ integration-tests/src/test/java/com/kenshoo/pl/audit/AuditForUpdateOneLevelTest.java
@@ -8,11 +8,12 @@ import com.kenshoo.pl.audit.commands.UpdateAuditedWithoutDataFieldsCommand;
 import com.kenshoo.pl.audit.commands.UpdateInclusiveAuditedCommand;
 import com.kenshoo.pl.audit.commands.UpdateNotAuditedCommand;
 import com.kenshoo.pl.entity.*;
+import com.kenshoo.pl.entity.audit.AuditRecord;
 import com.kenshoo.pl.entity.internal.audit.MainTable;
 import com.kenshoo.pl.entity.internal.audit.entitytypes.AuditedType;
 import com.kenshoo.pl.entity.internal.audit.entitytypes.AuditedWithoutDataFieldsType;
-import com.kenshoo.pl.entity.internal.audit.entitytypes.NotAuditedType;
 import com.kenshoo.pl.entity.internal.audit.entitytypes.InclusiveAuditedType;
+import com.kenshoo.pl.entity.internal.audit.entitytypes.NotAuditedType;
 import org.jooq.DSLContext;
 import org.junit.After;
 import org.junit.Before;
