+++ integration-tests/src/test/java/com/kenshoo/pl/audit/AuditForUpsertOneLevelTest.java
@@ -6,8 +6,9 @@ import com.kenshoo.jooq.TestJooqConfig;
 import com.kenshoo.pl.audit.commands.UpsertAuditedCommand;
 import com.kenshoo.pl.audit.commands.UpsertNotAuditedCommand;
 import com.kenshoo.pl.entity.*;
-import com.kenshoo.pl.entity.internal.audit.entitytypes.AuditedType;
+import com.kenshoo.pl.entity.audit.AuditRecord;
 import com.kenshoo.pl.entity.internal.audit.MainTable;
+import com.kenshoo.pl.entity.internal.audit.entitytypes.AuditedType;
 import com.kenshoo.pl.entity.internal.audit.entitytypes.NotAuditedType;
 import org.jooq.DSLContext;
 import org.junit.After;
