+++ main/src/main/java/com/kenshoo/pl/entity/PersistenceLayer.java
@@ -3,7 +3,11 @@ package com.kenshoo.pl.entity;
 import com.google.common.base.Stopwatch;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
-import com.kenshoo.pl.entity.internal.*;
+import com.kenshoo.pl.entity.audit.AuditRecord;
+import com.kenshoo.pl.entity.internal.ChangesFilter;
+import com.kenshoo.pl.entity.internal.EntitiesFetcher;
+import com.kenshoo.pl.entity.internal.EntitiesToContextFetcher;
+import com.kenshoo.pl.entity.internal.RequiredFieldsCommandsFilter;
 import com.kenshoo.pl.entity.internal.audit.RecursiveAuditRecordGenerator;
 import com.kenshoo.pl.entity.internal.validators.ValidationFilter;
 import com.kenshoo.pl.entity.spi.CurrentStateConsumer;
