+++ main/src/main/java/com/kenshoo/pl/entity/audit/AuditRecord.java
@@ -1,5 +1,7 @@
-package com.kenshoo.pl.entity;
+package com.kenshoo.pl.entity.audit;
 
+import com.kenshoo.pl.entity.ChangeOperation;
+import com.kenshoo.pl.entity.EntityType;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
