+++ main/src/main/java/com/kenshoo/pl/entity/internal/audit/RecursiveAuditRecordGenerator.java
@@ -1,6 +1,10 @@
 package com.kenshoo.pl.entity.internal.audit;
 
-import com.kenshoo.pl.entity.*;
+import com.kenshoo.pl.entity.ChangeContext;
+import com.kenshoo.pl.entity.ChangeFlowConfig;
+import com.kenshoo.pl.entity.EntityChange;
+import com.kenshoo.pl.entity.EntityType;
+import com.kenshoo.pl.entity.audit.AuditRecord;
 
 import java.util.Collection;
 import java.util.Optional;
