+++ main/src/test/java/com/kenshoo/pl/entity/internal/audit/entitytypes/AuditedChild2Type.java
@@ -1,11 +1,8 @@
 package com.kenshoo.pl.entity.internal.audit.entitytypes;
 
-import com.kenshoo.jooq.DataTable;
-import com.kenshoo.pl.entity.AbstractEntityType;
 import com.kenshoo.pl.entity.EntityField;
-import com.kenshoo.pl.entity.SupportedChangeOperation;
-import com.kenshoo.pl.entity.annotation.Audited;
 import com.kenshoo.pl.entity.annotation.Id;
+import com.kenshoo.pl.entity.annotation.audit.Audited;
 import com.kenshoo.pl.entity.internal.audit.ChildTable;
 
 @Audited
