+++ main/src/test/java/com/kenshoo/pl/entity/internal/audit/entitytypes/ExclusiveAuditedType.java
@@ -1,9 +1,9 @@
 package com.kenshoo.pl.entity.internal.audit.entitytypes;
 
 import com.kenshoo.pl.entity.EntityField;
-import com.kenshoo.pl.entity.annotation.Audited;
 import com.kenshoo.pl.entity.annotation.Id;
-import com.kenshoo.pl.entity.annotation.NotAudited;
+import com.kenshoo.pl.entity.annotation.audit.Audited;
+import com.kenshoo.pl.entity.annotation.audit.NotAudited;
 import com.kenshoo.pl.entity.internal.audit.MainTable;
 
 @Audited
