+++ main/src/test/java/com/kenshoo/pl/entity/matchers/audit/AuditRecordFieldRecordMatcher.java
@@ -1,8 +1,8 @@
 package com.kenshoo.pl.entity.matchers.audit;
 
-import com.kenshoo.pl.entity.AuditRecord;
 import com.kenshoo.pl.entity.EntityType;
-import com.kenshoo.pl.entity.FieldAuditRecord;
+import com.kenshoo.pl.entity.audit.AuditRecord;
+import com.kenshoo.pl.entity.audit.FieldAuditRecord;
 import org.hamcrest.Description;
 import org.hamcrest.TypeSafeMatcher;
