+++ main/src/test/java/com/kenshoo/pl/entity/matchers/audit/AuditRecordHasChildRecordMatcher.java
@@ -1,7 +1,7 @@
 package com.kenshoo.pl.entity.matchers.audit;
 
-import com.kenshoo.pl.entity.AuditRecord;
 import com.kenshoo.pl.entity.EntityType;
+import com.kenshoo.pl.entity.audit.AuditRecord;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
