+++ main/src/test/java/com/kenshoo/pl/entity/matchers/audit/AuditRecordOperatorMatcher.java
@@ -1,7 +1,7 @@
 package com.kenshoo.pl.entity.matchers.audit;
 
-import com.kenshoo.pl.entity.AuditRecord;
 import com.kenshoo.pl.entity.ChangeOperation;
+import com.kenshoo.pl.entity.audit.AuditRecord;
 import org.hamcrest.Description;
 import org.hamcrest.TypeSafeMatcher;
